Move ingredient add rows into their own subsection (#739)
